--- owner/src/main/java/org/the_chance/honeymart/ui/add_product/components/AddImageCard.kt
@@ -1,72 +1,0 @@
-package org.the_chance.honeymart.ui.add_product.components
-
-import android.net.Uri
-import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.activity.result.PickVisualMediaRequest
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.unit.dp
-import org.the_chance.design_system.R
-import org.the_chance.honymart.ui.theme.black60
-import org.the_chance.honymart.ui.theme.dimens
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun AddImageCard(
-    modifier: Modifier = Modifier
-) {
-    var selectedImageUris by remember {
-        mutableStateOf<List<Uri>>(emptyList())
-    }
-
-    val multiplePhotoPickerLauncher = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.PickMultipleVisualMedia(),
-        onResult = { uris -> selectedImageUris = uris }
-    )
-    Card(
-        modifier = modifier
-            .padding(
-                top = MaterialTheme.dimens.space16,
-                start = MaterialTheme.dimens.space16
-            )
-            .size(102.dp),
-        colors = CardDefaults.cardColors(MaterialTheme.colorScheme.tertiaryContainer),
-        onClick = {
-            multiplePhotoPickerLauncher.launch(
-                PickVisualMediaRequest(
-                    ActivityResultContracts.PickVisualMedia.ImageOnly
-                )
-            )
-        }
-    ) {
-        Box(
-            modifier = Modifier
-                .fillMaxSize()
-                .size(MaterialTheme.dimens.icon48),
-            contentAlignment = Alignment.Center
-        ) {
-            Icon(
-                painter = painterResource(R.drawable.icon_add_product),
-                contentDescription = "Icon Add",
-                modifier = Modifier.size(MaterialTheme.dimens.icon24),
-                tint = black60
-            )
-        }
-    }
-}
