+++ owner/src/main/java/org/the_chance/honeymart/ui/add_product/components/AddProductForm.kt
@@ -1,5 +1,10 @@
 package org.the_chance.honeymart.ui.add_product.components
 
+import android.net.Uri
+import androidx.activity.compose.rememberLauncherForActivityResult
+import androidx.activity.result.PickVisualMediaRequest
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -11,21 +16,28 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
+import androidx.compose.foundation.lazy.grid.GridCells
+import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
+import coil.compose.rememberAsyncImagePainter
 import org.the_chance.design_system.R
 import org.the_chance.honeymart.domain.util.ValidationState
 import org.the_chance.honeymart.ui.add_product.AddProductUiState
@@ -45,6 +57,15 @@ fun AddProductForm(
     onClickAddProduct: (name: String, price: Double, description: String) -> Unit,
     modifier: Modifier = Modifier
 ) {
+    var selectedImageUris by remember {
+        mutableStateOf<List<Uri>>(emptyList())
+    }
+
+    val multiplePhotoPickerLauncher = rememberLauncherForActivityResult(
+        contract = ActivityResultContracts.PickMultipleVisualMedia(4),
+        onResult = { uris -> selectedImageUris = uris }
+    )
+
     Column(
         modifier = modifier
             .fillMaxSize()
@@ -52,7 +73,6 @@ fun AddProductForm(
                 color = MaterialTheme.colorScheme.tertiary,
                 shape = MaterialTheme.shapes.medium
             )
-            .verticalScroll(rememberScrollState()),
     ) {
         Row(
             modifier = Modifier.padding(
@@ -121,28 +141,56 @@ fun AddProductForm(
             color = MaterialTheme.colorScheme.onSecondaryContainer,
             textAlign = TextAlign.Center,
         )
-        Card(
+        Row(
             modifier = Modifier
-                .padding(
-                    top = MaterialTheme.dimens.space16,
-                    start = MaterialTheme.dimens.space16
-                )
-                .size(102.dp),
-            colors = CardDefaults.cardColors(MaterialTheme.colorScheme.tertiaryContainer),
-            onClick = { }
+                .fillMaxWidth()
+                .padding(MaterialTheme.dimens.space16)
         ) {
-            Box(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .size(MaterialTheme.dimens.icon48),
-                contentAlignment = Alignment.Center
-            ) {
-                Icon(
-                    painter = painterResource(R.drawable.icon_add_product),
-                    contentDescription = "Icon Cart",
-                    modifier = Modifier.size(MaterialTheme.dimens.icon24),
-                    tint = black60
-                )
+            LazyVerticalGrid(columns = GridCells.Adaptive(102.dp)) {
+                items(selectedImageUris.size + 1) { index ->
+                    if (index < selectedImageUris.size) {
+                        Surface(
+                            shape = MaterialTheme.shapes.medium,
+                            color = MaterialTheme.colorScheme.tertiaryContainer,
+                            modifier = Modifier
+                                .size(102.dp)
+                                .padding(4.dp)
+                        ) {
+                            Image(
+                                painter = rememberAsyncImagePainter(selectedImageUris[index]),
+                                contentScale = ContentScale.Crop,
+                                contentDescription = "Image of product",
+                            )
+                        }
+                    } else if (index <= 3) {
+                        Card(
+                            modifier = Modifier
+                                .size(102.dp)
+                                .padding(4.dp),
+                            colors = CardDefaults.cardColors(MaterialTheme.colorScheme.tertiaryContainer),
+                            onClick = {
+                                multiplePhotoPickerLauncher.launch(
+                                    PickVisualMediaRequest(ActivityResultContracts.PickVisualMedia.ImageOnly)
+                                )
+                            },
+                            shape = MaterialTheme.shapes.medium
+                        ) {
+                            Box(
+                                modifier = Modifier
+                                    .fillMaxSize()
+                                    .size(MaterialTheme.dimens.icon48),
+                                contentAlignment = Alignment.Center
+                            ) {
+                                Icon(
+                                    painter = painterResource(R.drawable.icon_add_product),
+                                    contentDescription = "Icon Add",
+                                    modifier = Modifier.size(MaterialTheme.dimens.icon24),
+                                    tint = black60
+                                )
+                            }
+                        }
+                    }
+                }
             }
         }
         Spacer(modifier = Modifier.weight(1F))
